part 2 of review cleaning
